Set RestTemplate error handler when instantiating it
